refactor: code quality and architecture overhaul
Architecture:
- Introduce unexported operations interface (Strategy pattern) with
  v1Operations and v2Operations implementations
- Extend KVStorage interface for dependency injection (Pipeline, SetNX,
  Publish, Subscribe, Close, etc.)
- Wire redisStorage adapter in Create() - all Redis ops through KVStorage

Error handling:
- Activate structured error helpers (wrapV1Err/wrapV2Err) in v1 and v2 ops
- Add ErrCacheRequiresV2 guard for cache operations
- Add error wrapping tests for both V1 and V2 operations

Context propagation:
- Add 13 public *Context methods (AddContext, RemoveContext, FindContext, etc.)
- Deprecate old methods with godoc comments, delegating to *Context variants
- Propagate context through batch and parallel helpers

Code quality:
- Rename underscore-prefixed methods in trie.go for clarity
- Add BatchOptions to RemoveMany for consistency with AddMany
- Replace direct redisClient calls with storage interface methods
  in cache.go, pubsub.go, trie.go
- Fix errcheck warnings on pipeline command return values
- Fix variable shadowing in v1Operations.remove()
- Extract magic number for pubsub channel buffer size

Testing:
- Split monolithic acor_test.go (2000+ lines) into feature-based files
- Update all tests for DI+Strategy restructuring
- Add V1/V2 error wrapping test suites

CI/Docs/Community:
- Add fuzz testing step to CI pipeline
- Fix API method names in docs/ (Hugo content)
- Add issue/PR templates and SECURITY.md
- Add new Makefile targets and update .gitignore

No public API signature changes. No new dependencies. Pure refactoring.
