+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,38 @@
+---
+name: Bug Report
+about: Report a bug to help us improve ACOR
+title: '[BUG] '
+labels: bug
+assignees: ''
+---
+
+## Description
+
+<!-- A clear and concise description of the bug -->
+
+## Environment
+
+- **Go version:**
+- **Redis version:**
+- **OS:**
+- **ACOR version:**
+
+## Steps to Reproduce
+
+<!-- Steps to reproduce the bug -->
+
+1.
+2.
+3.
+
+## Expected Behavior
+
+<!-- What you expected to happen -->
+
+## Actual Behavior
+
+<!-- What actually happened -->
+
+## Additional Context
+
+<!-- Any other context about the problem, such as code snippets, stack traces, etc. -->
