+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,0 +1,5 @@
+blank_issues_enabled: false
+contact_links:
+  - name: Documentation
+    url: https://skyoo2003.github.io/acor/
+    about: Check the documentation before opening an issue.
