+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,23 @@
+---
+name: Feature Request
+about: Suggest a new feature or enhancement for ACOR
+title: '[FEATURE] '
+labels: enhancement
+assignees: ''
+---
+
+## Problem Description
+
+<!-- A clear and concise description of the problem this feature would solve -->
+
+## Proposed Solution
+
+<!-- A clear and concise description of the proposed feature -->
+
+## Alternatives Considered
+
+<!-- Any alternative solutions or features you've considered -->
+
+## Additional Context
+
+<!-- Any other context, screenshots, or mockups about the feature request -->
