+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,28 @@
+## Description
+
+<!-- Describe your changes and the purpose of this PR -->
+
+## Type of Change
+
+<!-- What type of change does this PR introduce? (Check all that apply) -->
+
+- [ ] Bug fix (non-breaking change which fixes an issue)
+- [ ] New feature (non-breaking change which adds functionality)
+- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
+- [ ] Documentation update
+- [ ] Refactoring (no functional changes)
+- [ ] Test update
+
+## Checklist
+
+<!-- Ensure your PR meets the following requirements -->
+
+- [ ] Tests pass (`make test`)
+- [ ] Linting passes (`make lint`)
+- [ ] Build succeeds (`make build`)
+- [ ] Documentation updated if needed
+- [ ] Commit messages follow guidelines
+
+## Additional Notes
+
+<!-- Any other information reviewers should know -->
