+++ Makefile
@@ -1,6 +1,6 @@
-.PHONY: all setup clean build test lint coverage
+.PHONY: all setup clean build test lint lint-fix coverage vet fuzz race
 
-all: lint test build
+all: vet lint test build
 
 setup:
 	@pre-commit install
@@ -22,3 +22,15 @@ coverage:
 	@go test ./... -coverprofile=coverage.out -covermode=atomic
 	@go tool cover -html=coverage.out -o coverage.html
 	@echo "Coverage report: coverage.html"
+
+vet:
+	@go vet ./...
+
+lint-fix:
+	@golangci-lint run --fix ./...
+
+fuzz:
+	@go test -fuzz=Fuzz -fuzztime=30s ./pkg/acor/...
+
+race:
+	@go test -race ./...
