+++ SECURITY.md
@@ -1,0 +1,46 @@
+# Security Policy
+
+## Supported Versions
+
+The following versions of ACOR are currently supported with security updates:
+
+| Version | Supported          |
+| ------- | ------------------ |
+| latest  | :white_check_mark: |
+
+## Reporting a Vulnerability
+
+If you discover a security vulnerability within ACOR, please report it responsibly.
+
+**Please do not report security vulnerabilities through public GitHub issues.**
+
+Instead, please report them privately through [GitHub Security Advisories](https://github.com/skyoo2003/acor/security/advisories/new).
+
+When reporting, please include:
+
+- A description of the vulnerability
+- Steps to reproduce the issue
+- Potential impact of the vulnerability
+- Any suggested fixes (if applicable)
+
+## Response Timeline
+
+We aim to acknowledge vulnerability reports within 48 hours and provide a timeline for fixes based on severity:
+
+- **Critical**: Fix within 7 days
+- **High**: Fix within 30 days
+- **Medium**: Fix within 90 days
+- **Low**: Fix in next release
+
+## Disclosure Policy
+
+We follow a coordinated disclosure process and will work with reporters to:
+
+1. Confirm the vulnerability
+2. Develop and test a fix
+3. Release a patched version
+4. Credit reporters (unless anonymity is requested)
+
+## License
+
+This security policy is part of the [MIT License](LICENSE) project.
