+++ pkg/acor/batch_test.go
@@ -5,6 +5,8 @@ import (
 	"testing"
 )
 
+const testKeywordHim = "him"
+
 func TestAddManyBestEffort(t *testing.T) {
 	ac, mr := createAhoCorasick(t)
 	defer mr.Close()
@@ -159,7 +161,7 @@ func TestRemoveMany(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	result, err := ac.RemoveMany([]string{"he", "her"})
+	result, err := ac.RemoveMany([]string{"he", "her"}, nil)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -168,11 +170,11 @@ func TestRemoveMany(t *testing.T) {
 		t.Errorf("expected 2 removed, got %d", len(result.Removed))
 	}
 
-	results, err := ac.Find("him")
+	results, err := ac.Find(testKeywordHim)
 	if err != nil {
 		t.Fatal(err)
 	}
-	if len(results) != 1 || results[0] != "him" {
+	if len(results) != 1 || results[0] != testKeywordHim {
 		t.Error("expected 'him' to remain")
 	}
 }
@@ -187,7 +189,7 @@ func TestRemoveManyWithDuplicates(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	result, err := ac.RemoveMany([]string{"he", "he", "her"})
+	result, err := ac.RemoveMany([]string{"he", "he", "her"}, nil)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -197,6 +199,109 @@ func TestRemoveManyWithDuplicates(t *testing.T) {
 	}
 	if len(result.Skipped) != 1 {
 		t.Errorf("expected 1 skipped duplicate, got %d", len(result.Skipped))
+	}
+}
+
+func TestRemoveManyBestEffort(t *testing.T) {
+	ac, mr := createAhoCorasick(t)
+	defer mr.Close()
+	defer func() { _ = ac.Close() }()
+	defer func() { _ = ac.Flush() }()
+
+	if _, err := ac.AddMany([]string{"he", "her", "him"}, nil); err != nil {
+		t.Fatal(err)
+	}
+
+	result, err := ac.RemoveMany([]string{"he", "", "him"}, &BatchOptions{Mode: BatchModeBestEffort})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if len(result.Removed) != 2 {
+		t.Errorf("expected 2 removed, got %d", len(result.Removed))
+	}
+	if len(result.Failed) != 1 {
+		t.Errorf("expected 1 failed, got %d", len(result.Failed))
+	}
+	if result.Failed[0].Keyword != "" {
+		t.Errorf("expected failed keyword '', got %q", result.Failed[0].Keyword)
+	}
+	if !errors.Is(result.Failed[0].Error, ErrEmptyKeyword) {
+		t.Errorf("expected ErrEmptyKeyword, got %v", result.Failed[0].Error)
+	}
+}
+
+func TestRemoveManyNilOpts(t *testing.T) {
+	ac, mr := createAhoCorasick(t)
+	defer mr.Close()
+	defer func() { _ = ac.Close() }()
+	defer func() { _ = ac.Flush() }()
+
+	if _, err := ac.AddMany([]string{"he", "her", "him"}, nil); err != nil {
+		t.Fatal(err)
+	}
+
+	result, err := ac.RemoveMany([]string{"he", "her"}, nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if len(result.Removed) != 2 {
+		t.Errorf("expected 2 removed, got %d", len(result.Removed))
+	}
+}
+
+func TestRemoveManyTransactional(t *testing.T) {
+	ac, mr := createAhoCorasick(t)
+	defer mr.Close()
+	defer func() { _ = ac.Close() }()
+	defer func() { _ = ac.Flush() }()
+
+	if _, err := ac.AddMany([]string{"he", "her", "him"}, nil); err != nil {
+		t.Fatal(err)
+	}
+
+	result, err := ac.RemoveMany([]string{"he", "her"}, &BatchOptions{Mode: BatchModeTransactional})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if len(result.Removed) != 2 {
+		t.Errorf("expected 2 removed, got %d", len(result.Removed))
+	}
+
+	results, err := ac.Find(testKeywordHim)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(results) != 1 || results[0] != testKeywordHim {
+		t.Error("expected 'him' to remain")
+	}
+}
+
+func TestRemoveManyTransactionalRollbackOnError(t *testing.T) {
+	ac, mr := createAhoCorasick(t)
+	defer mr.Close()
+	defer func() { _ = ac.Close() }()
+	defer func() { _ = ac.Flush() }()
+
+	if _, err := ac.AddMany([]string{"he", "her", "him"}, nil); err != nil {
+		t.Fatal(err)
+	}
+
+	_, err := ac.RemoveMany([]string{"he", "", "him"}, &BatchOptions{Mode: BatchModeTransactional})
+	if !errors.Is(err, ErrEmptyKeyword) {
+		t.Fatalf("expected ErrEmptyKeyword, got %v", err)
+	}
+
+	for _, kw := range []string{"he", "her", "him"} {
+		results, findErr := ac.Find(kw)
+		if findErr != nil {
+			t.Fatal(findErr)
+		}
+		if len(results) == 0 {
+			t.Errorf("expected rollback to restore %q, but it was not found", kw)
+		}
 	}
 }
